+++ sdk/src/main/java/com/hedera/hashgraph/sdk/TransactionResponse.java
@@ -47,11 +47,11 @@ public final class TransactionResponse {
 
     public final TransactionId transactionId;
 
-    private boolean validateStatus = true;
-
     @Nullable
     @Deprecated
     public final TransactionId scheduledTransactionId;
+
+    private boolean validateStatus = true;
 
     /**
      * Constructor.
@@ -115,11 +115,9 @@ public final class TransactionResponse {
      * @throws ReceiptStatusException       when there is an issue with the receipt
      */
     public TransactionReceipt getReceipt(Client client, Duration timeout) throws TimeoutException, PrecheckStatusException, ReceiptStatusException {
-        var receipt = getReceiptQuery().execute(client, timeout);
-
-        if (validateStatus && receipt.status != Status.SUCCESS) {
-            throw new ReceiptStatusException(transactionId, receipt);
-        }
+        var receipt = getReceiptQuery()
+            .execute(client, timeout)
+            .validateStatus(validateStatus);
 
         return receipt;
     }
@@ -148,7 +146,15 @@ public final class TransactionResponse {
      * @return                          the transaction receipt
      */
     public CompletableFuture<TransactionReceipt> getReceiptAsync(Client client, Duration timeout) {
-        return getReceiptQuery().executeAsync(client, timeout);
+        return getReceiptQuery()
+            .executeAsync(client, timeout)
+            .thenCompose(receipt -> {
+                try {
+                    return CompletableFuture.completedFuture(receipt.validateStatus(validateStatus));
+                } catch (ReceiptStatusException e) {
+                    return CompletableFuture.failedFuture(e);
+                }
+            });
     }
 
     /**