+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
  - `TransactionResponse.[get|set]ValidateStatus()`
+ - `TransactionReceipt.validateStatus()`
+ - `TransactionRecord.validateReceiptStatus()`
+ - `TransactionReceipt.transactionId`
  - `TopicUpdateTransaction.[get|set]ExpirationTime()`
  - `CustomFee.[set|get]AllCollectorsAreExempt()`